refactor(responsive appbar): refactored mobile appbar into its own component for readability
