+++ src/layout/MobileAppbar.component.jsx
@@ -1,0 +1,60 @@
+import { Box, IconButton, Typography, Menu, MenuItem } from "@mui/material";
+import MenuIcon from "@mui/icons-material/Menu";
+import { menuItems } from "../routesAndMenutItems";
+import { useState } from "react";
+import { Link } from "react-router-dom";
+
+export function MobileAppBar() {
+  const [anchorElNav, setAnchorElNav] = useState(null);
+
+  const handleOpenNavMenu = (event) => {
+    setAnchorElNav(event.currentTarget);
+  };
+
+  const handleCloseNavMenu = () => {
+    setAnchorElNav(null);
+  };
+  return (
+    <Box sx={{ flexGrow: 1, display: { xs: "flex", md: "none" } }}>
+      <IconButton
+        size="large"
+        aria-label="account of current user"
+        aria-controls="menu-appbar"
+        aria-haspopup="true"
+        onClick={handleOpenNavMenu}
+        color="inherit"
+      >
+        <MenuIcon />
+      </IconButton>
+      <Menu
+        id="menu-appbar"
+        anchorEl={anchorElNav}
+        anchorOrigin={{
+          vertical: "bottom",
+          horizontal: "left",
+        }}
+        keepMounted
+        transformOrigin={{
+          vertical: "top",
+          horizontal: "left",
+        }}
+        open={Boolean(anchorElNav)}
+        onClose={handleCloseNavMenu}
+        sx={{
+          display: { xs: "block", md: "none" },
+        }}
+      >
+        {menuItems.map((page) => (
+          <MenuItem
+            key={page.title}
+            component={Link}
+            onClick={handleCloseNavMenu}
+            to={page.link}
+          >
+            <Typography textAlign="center">{page.title}</Typography>
+          </MenuItem>
+        ))}
+      </Menu>
+    </Box>
+  );
+}
