+++ src/layout/ResponsiveAppBar.component.jsx
@@ -1,82 +1,18 @@
-import { useState } from "react";
-import {
-  AppBar,
-  Box,
-  Toolbar,
-  IconButton,
-  Typography,
-  Menu,
-  Container,
-  Button,
-  MenuItem,
-} from "@mui/material";
-
-import MenuIcon from "@mui/icons-material/Menu";
-import { useNavigate } from "react-router-dom";
+import { AppBar, Box, Toolbar, Container, Button } from "@mui/material";
 import { menuItems } from "../routesAndMenutItems";
+import { MobileAppBar } from "./MobileAppbar.component";
 
 export function ResponsiveAppBar() {
-  const [anchorElNav, setAnchorElNav] = useState(null);
-
-  const history = useNavigate();
-  const handleOpenNavMenu = (event) => {
-    setAnchorElNav(event.currentTarget);
-  };
-
-  const handleCloseNavMenu = (link) => {
-    setAnchorElNav(null);
-    history(`${link}`);
-  };
-
   return (
     <AppBar position="static">
       <Container maxWidth="xl">
         <Toolbar disableGutters>
-          <Box sx={{ flexGrow: 1, display: { xs: "flex", md: "none" } }}>
-            <IconButton
-              size="large"
-              aria-label="account of current user"
-              aria-controls="menu-appbar"
-              aria-haspopup="true"
-              onClick={handleOpenNavMenu}
-              color="inherit"
-            >
-              <MenuIcon />
-            </IconButton>
-            <Menu
-              id="menu-appbar"
-              anchorEl={anchorElNav}
-              anchorOrigin={{
-                vertical: "bottom",
-                horizontal: "left",
-              }}
-              keepMounted
-              transformOrigin={{
-                vertical: "top",
-                horizontal: "left",
-              }}
-              open={Boolean(anchorElNav)}
-              onClose={handleCloseNavMenu}
-              sx={{
-                display: { xs: "block", md: "none" },
-              }}
-            >
-              {menuItems.map((page) => (
-                <MenuItem
-                  key={page.title}
-                  onClick={() => handleCloseNavMenu(page.link)}
-                >
-                  <Typography textAlign="center">{page.title}</Typography>
-                </MenuItem>
-              ))}
-            </Menu>
-          </Box>
-
+          <MobileAppBar />
           <Box sx={{ display: { xs: "none", md: "flex" } }}>
             {menuItems.map((page) => (
               <Button
                 key={page.title}
-                onClick={handleCloseNavMenu}
+                // onClick={handleCloseNavMenu}
                 sx={{ my: 2, color: "white", display: "block" }}
                 href={page.link}
               >
