+++ src/layout/index.js
@@ -1,1 +1,2 @@
 export * from "./ResponsiveAppBar.component";
+export * from "./MobileAppbar.component";
